requirements.txt
boto3==1.26.137
python-dotenv==1.0.0
pyspark==3.5.0
pyathena==3.10.3
requests==2.31.0
psycopg2-binary==2.9.6
streamlit==1.28.1
plotly==5.17.0
pandas==2.0.3